requirements.txt
Flask
tqdm
yacs
h5py
numpy
pandas
scipy
numba
smplx
gdown
PyYAML
joblib
pillow
trimesh
pyrender
progress
filterpy
matplotlib
tensorflow
tensorboard
torchvision
scikit-image
scikit-video
opencv-python
llvmlite
torch-dct
git+https://github.com/mattloper/chumpy.git
git+https://github.com/giacaglia/pytube.git
git+https://github.com/mkocabas/yolov3-pytorch.git
git+https://github.com/mkocabas/multi-person-tracker.git
